requirements.txt
Django==1.8.3
celery==3.1.18
djangorestframework==3.1.3
librabbitmq==1.6.1
redis==2.10.3
uWSGI==2.0.11.1
